docker/agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
litellm==1.51.0
pydantic==2.9.2
psycopg2-binary==2.9.9
prometheus-client==0.20.0
httpx==0.27.2
python-dotenv==1.0.1
